Component back to being a class! Need it as such for reflection or something...
ComponentProvider is no longer an interface, but instead a proper self contained class! With this, a new rule is enforced that components must support the default constructor (no parameters). This way, a platform simply provides a map from core type to implementation type.

New project added for the Unity platform implementation!