Adds some nice ways of dealing with unconstrained space
- "DeferredSize" is a new widget which, like Flex, does nothing except
provide info to other widget(s). In particular, when used as a direct
child of Stack, it will take the size of the stack after other widgets
have layed out.
  e.g: A ColoredBox could be made to match the size of a widget stacked
  in front of it.

- In a SequenceLayout, an index can be specified to indicate a single
child widget which will be used to constrain the cross axis.
  e.g: A column could be made to be the width of a title.

- Images can now have one axis unconstrained, where the returned size
will be the ratio of the image, with a side length of the constrained
axis.