Changes how widget replacement works.
Before, the position of the child had to be the same, as well as passing
the "CanReplace" test, which was an overridable method of Widget.

CanReplace being virtual adds unneeded complexity for a user, and
forcing the position to be the same could be quite limiting.

So instead, I've mimiced Flutter, which I probably should have done in
the first place...

Widgets can now set a "key," which is just a string, and widgets will
replace another with the exact same type and key, regardless of
position. If more than one meets the predicate, the first will be taken.