Big'ol rewrite!
The whole build system has been rewritten, to make some sense.

Previously, when we did build operations, we make sure that at every
step each tree was maintained in working order. This allowed us to do
the operation "in place" keeping just one set of trees (widget, element,
and input).

The goal there was sound: focus all the logic into individule nodes,
allowing everything to be optimally recursive. And it worked, and
created an initially easy mental model. The issue, though, came out of
this eased, recursive mental model: we were actually doing a lot of up
and down search operations at every node.

So in the end, we did a lot more work than necessary. (There was also a
bug where we did exponentially more work than necessary, but we don't
talk about that...)

In particular, when doing a rebuild, we end up tearing apart the entire
sub-tree, at each step.

Realizing this, the new model simplifies things dramatically by simply
building a whole new tree everytime an update is required. It does this
referring to the previous build, copying data where it can.

Additionally, we throw out the idea of building an "element" and input
tree at the same time. Doing so just adds extra complexity in trying to
manage the other subtrees, propagating the same issues.

Now, after building the widget tree, we go through and directly update
changed "elements" (now called DisplayBox), and build the input tree
from scratch.

Performance wise, this should be about the same for initial builds
(maybe a little bit slower), and much faster for the average rebuild,
and exponentially faster for some particularly nasty rebuilds.

Code complexity wise, this is a massive boon.